requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==1.5.3
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.5.0
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
jinja2==3.1.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1